+++ lending/src/codatlending/financial_statements.py
@@ -170,7 +170,7 @@ class FinancialStatements:
     
     def get_categorized_balance_sheet(self, request: operations.GetCategorizedBalanceSheetStatementRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetCategorizedBalanceSheetStatementResponse:
         r"""Get categorized balance sheet statement
-        The *Get enhanced balance sheet accounts* endpoint returns a list of categorized accounts that appear on a company’s Balance Sheet along with a balance per financial statement date.
+        The *Get categorized balance sheet statement* endpoint returns a list of categorized accounts that appear on a company’s Balance Sheet along with a balance per financial statement date.
 
         Codat suggests a category for each account automatically, but you can [change it](https://docs.codat.io/lending/enhanced-financials/overview#categorize-accounts) to a more suitable one.
         """
@@ -222,7 +222,7 @@ class FinancialStatements:
     
     def get_categorized_profit_and_loss(self, request: operations.GetCategorizedProfitAndLossStatementRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetCategorizedProfitAndLossStatementResponse:
         r"""Get categorized profit and loss statement
-        The *Get enhanced profit and loss accounts* endpoint returns a list of categorized accounts that appear on a company’s Profit and Loss statement. It also includes a balance as of the financial statement date.
+        The *Get categorized profit and loss statement* endpoint returns a list of categorized accounts that appear on a company’s Profit and Loss statement. It also includes a balance as of the financial statement date.
 
         Codat suggests a category for each account automatically, but you can [change it](https://docs.codat.io/lending/enhanced-financials/overview#categorize-accounts) to a more suitable one.
         """