+++ lending/src/codatlending/liabilities.py
@@ -15,7 +15,7 @@ class Liabilities:
     
     def get_loan_summary(self, request: operations.GetLoanSummaryRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetLoanSummaryResponse:
         r"""Get loan summaries
-        The *Get enhanced loan summaries* endpoint returns a summary by integration type of all loans identified in each integration.
+        The *Get loan summaries* endpoint returns a summary by integration type of all loans identified in each integration.
 
         The endpoint returns a list of a company's [loan summaries](https://docs.codat.io/lending-api#/schemas/LoanSummary) for each valid data connection.
 
@@ -68,7 +68,7 @@ class Liabilities:
     
     def list_loan_transactions(self, request: operations.ListLoanTransactionsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.ListLoanTransactionsResponse:
         r"""List loan transactions
-        The *List enhanced loan transactions* endpoint returns all [loan transactions](https://docs.codat.io/lending-api#/schemas/LoanTransactions) identified from a company's accounting, banking, and commerce integrations.
+        The *List loan transactions* endpoint returns all [loan transactions](https://docs.codat.io/lending-api#/schemas/LoanTransactions) identified from a company's accounting, banking, and commerce integrations.
 
         This detail gives analysts a better idea of the loan obligations a company may have.