+++ lending/src/codatlending/sales.py
@@ -82,7 +82,7 @@ class Sales:
         - __Retention rate__: the ratio of existing customers within the specified period compared to the total customers at the end of the previous period represented as a percentage.
         - __Repeat rate__: the ratio of existing customers to total customers over the specified period represented as a percentage.
 
-        Learn more about the formulas used to calculate customer retention metrics [here](/lending/commerce-metrics/overview#what-metrics-are-available).
+        Learn more about the formulas used to calculate customer retention metrics [here](https://docs.codat.io/lending/commerce-metrics/overview#what-metrics-are-available).
 
         #### Response structure
 
@@ -207,9 +207,9 @@ class Sales:
 
         This detail helps you assess a merchant's health and advise them on performance improvement strategies. It also provides you with key insights you need to assess the credit risk of a company.
 
-        Learn more about the formulas used to calculate the lifetime value metrics [here](/lending/commerce-metrics/overview#what-metrics-are-available).
+        Learn more about the formulas used to calculate the lifetime value metrics [here](https://docs.codat.io/lending/commerce-metrics/overview#what-metrics-are-available).
 
-        Refer to the [commerce reporting structure](/lending/commerce-metrics/reporting-structure) page for more detail on commerce reports in Lending.
+        Refer to the [commerce reporting structure](https://docs.codat.io/lending/commerce-metrics/reporting-structure) page for more detail on commerce reports in Lending.
 
         #### Response structure
 
@@ -388,9 +388,9 @@ class Sales:
 
         This detail helps you assess a merchant's health and advise them on performance improvement strategies. It also provides you with key insights you need to assess the credit risk of a company. 
 
-        Learn more about the formulas used to calculate the order metrics [here](/lending/commerce-metrics/overview#what-metrics-are-available).
+        Learn more about the formulas used to calculate the order metrics [here](https://docs.codat.io/lending/commerce-metrics/overview#what-metrics-are-available).
 
-        Refer to the [commerce reporting structure](/lending/commerce-metrics/reporting-structure) page for more details on commerce reports in Lending.
+        Refer to the [commerce reporting structure](https://docs.codat.io/lending/commerce-metrics/reporting-structure) page for more details on commerce reports in Lending.
 
         #### Response structure
 
@@ -680,9 +680,9 @@ class Sales:
 
         This detail helps you assess a merchant's health and advise them on performance improvement strategies. It also provides you with key insights you need to assess the credit risk of a company. 
 
-        Learn more about the formulas used to calculate the refunds metrics [here](/lending/commerce-metrics/overview#what-metrics-are-available).
+        Learn more about the formulas used to calculate the refunds metrics [here](https://docs.codat.io/lending/commerce-metrics/overview#what-metrics-are-available).
 
-        Refer to the [commerce reporting structure](/lending/commerce-metrics/reporting-structure) page for more details on commerce reports in Lending.
+        Refer to the [commerce reporting structure](https://docs.codat.io/lending/commerce-metrics/reporting-structure) page for more details on commerce reports in Lending.
 
         #### Response structure
 
@@ -749,7 +749,29 @@ class Sales:
     
     def get_revenue_metrics(self, request: operations.GetCommerceRevenueMetricsRequest, retries: Optional[utils.RetryConfig] = None) -> operations.GetCommerceRevenueMetricsResponse:
         r"""Get commerce revenue metrics
-        Get the revenue and revenue growth for a specific company connection, over one or more periods of time.
+        The *Get revenue report* endpoint returns the revenue and revenue growth for a specific company connection over one or more periods of time.
+
+        This detail helps you assess a merchant's health and advise them on performance improvement strategies. It also provides you with key insights you need to assess the credit risk of a company. 
+
+        Learn more about the formulas used to calculate the revenue metrics [here](https://docs.codat.io/lending/commerce-metrics/overview#what-metrics-are-available).
+
+        Refer to the [commerce reporting structure](https://docs.codat.io/lending/commerce-metrics/reporting-structure) page for more details on commerce reports in Lending.
+
+        #### Response structure
+
+        The Revenue report's dimensions and measures are:
+
+        | Index         | Dimensions |
+        |---------------|------------|
+        |   `index` = 0 | Period     |
+        |   `index` = 1 | Revenue    |
+
+        | Index         | Measures                                                                                                                 |
+        |---------------|--------------------------------------------------------------------------------------------------------------------------|
+        | `index` = 0   | Value                                                                                                                    |
+        | `index` = 1   | Percentage change, defined as the change between the current and previous periods' values and expressed as a percentage. |
+
+        The report data then combines multiple reporting dimensions and measures and outputs the value of each combination.
         """
         base_url = utils.template_url(*self.sdk_configuration.get_server_details())