+++ lending/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.2.0] lending
 ### Releases
 - [PyPI v0.2.0] https://pypi.org/project/codat-lending/0.2.0 - lending
+
+## 2023-09-09 00:15:47
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.82.3 (2.107.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.2.1] lending
+### Releases
+- [PyPI v0.2.1] https://pypi.org/project/codat-lending/0.2.1 - lending